refactor: move builtins to libs and add semantic checks
