+++ libs/mql-interpreter/src/libs/builtins/iMACD.ts
