+++ libs/mql-interpreter/src/libs/builtins/math.ts
