+++ libs/mql-interpreter/src/libs/builtins/strings.ts
