+++ libs/mql-interpreter/src/parser/parser.ts
@@ -1,7 +1,6 @@
 import { Token, TokenType } from "./lexer";
 import {
   EnumMember,
-  SourceLocation,
   EnumDeclaration,
   ClassField,
   ClassMethod,
