+++ libs/mql-interpreter/test/runtime/builtins/file.test.ts
@@ -1,12 +1,12 @@
 import { describe, it, expect, beforeEach } from "vitest";
 import { VirtualTerminal } from "../../../src/libs/virtualTerminal";
-import { setTerminal } from "../../../src/runtime/builtins/common";
+import { setTerminal } from "../../../src/libs/builtins/common";
 import {
   FileOpen,
   FileReadString,
   FileWriteString,
   FileClose,
-} from "../../../src/runtime/builtins/file";
+} from "../../../src/libs/builtins/file";
 
 describe("file builtins", () => {
   beforeEach(() => {
